Full working Rule Manager and Traptor
Traptor now fully works with Rule Manager. Add a rule to Cooper and
Rule Manager picks up on it, adds the rule to Redis, and informs the
appropriate Traptor. No Traptor of that type? No problem. Traptor
handles that too. Once the Traptor sees a message for itself it
restarts, thus refreshing the rules it uses. A Traptor will restart
only if it finds a rule for itself. No others need apply.