Update Traptor find rule match logic and rule handling
Make all the things unicode!

Much better handling of unicode is now in place. The redis connection
automatically decodes responses. In addition, the _find_rule_matches
now checks all keys and sub-keys of tweet objects to add a rule match
for the tweet. I found that the location of the value from a rule could
be in a key or a child key. All cases are now being handled.

And if that wasn’t enough, Traptor can now handle comma-separated rules.

I also commented out some debug statements, and moved the restart_flag
check so that Traptor doesn’t have to wait to process an item in order
to check to see if it needs to restart.