Add heartbeat to Redis every 5 seconds
This allows anything with access to the same Redis database to know if
a Traptor is alive, or not.

Adds an expiring key to Redis every 5 seconds that contains the traptor
type, traptor id, and timestamp of the key. If the key exists it’ll be
updated. If the Traptor dies the it will go away.