Update stats to use the rule_id field
Updated the stats collection to use the rule_id for the key. We use
this in the rule creation from rule manager. Having it as the stats key
will allow us to compare on keys rather than needing to retrieve all
the data from Redis in order to perform comparisons.